Add fault tolerance, dropped-scene reporting, and bounds CRS clip
- Bounds mode: when input_crs differs from target_crs, synthesise a
  densified polygon from the input rectangle so per-scene masks clip
  valid pixels to the requested area. transform_bounds previously gave
  an axis-aligned UTM envelope that overshoots the requested lat/lon
  rectangle by tens of km on wide AOIs (e.g. WA strip in EPSG:32751).
  Sidecar mode still reflects the user's request. Regression test
  included.
- Tile aggregation: requeue specs that fail with RasterioIOError to the
  back of the executor queue. Per-tile cap (MAX_REQUEUES_PER_TILE) and
  whole-run cap (MAX_TOTAL_REQUEUE_FRACTION) keep systemic failures
  surfacing fast.
- Shared exponential backoff_delay helper with cap + jitter, used by
  the per-scene fetch decorator and the per-tile retry path. Avoids
  thundering-herd retries when many workers back off in lockstep.
- Surface dropped scenes after Phase 1: report_dropped_scenes prints to
  stderr (visible without a logging handler) and the same list is
  embedded in the sidecar metadata for grid and bounds pipelines.
- gdal_env: enable curl-layer HTTP retry (GDAL_HTTP_MAX_RETRY=3,
  RETRY_DELAY=1) so transient 429/5xx and TLS hiccups don't burn the
  app-level retry budget in a few seconds.
- Default ocm_inference_dtype to "fp32" — runs on every backend
  out-of-the-box and is the fastest CPU option. GPU users opt in to
  fp16/bf16 for speedups. CHANGELOG, README, and mosaic() docstring
  updated.
- README: document pytest/ruff commands and the slow-test marker.

+++ s2mosaic/config.py
@@ -106,7 +106,7 @@ class MosaicRequest:
     scene_sort_fn: Optional[Callable[..., Any]] = None
     cloud_mask: str = CLOUD_MASK_OCM
     ocm_batch_size: int = 1
-    ocm_inference_dtype: str = "fp16"
+    ocm_inference_dtype: str = "fp32"
     output_crs: Optional[int] = None
     resolution: int = 10
     resampling_method: str = "nearest"
